+++ Core/Domains/Instrument.cs
@@ -14,32 +14,32 @@ namespace Terminal.Core.Domains
     /// <summary>
     /// Long swap rate for keeping position overnight
     /// </summary>
-    double? SwapLong { get; set; }
+    decimal? SwapLong { get; set; }
 
     /// <summary>
     /// Short swap rate for keeping position overnight
     /// </summary>
-    double? SwapShort { get; set; }
+    decimal? SwapShort { get; set; }
 
     /// <summary>
     /// Commission
     /// </summary>
-    double? Commission { get; set; }
+    decimal? Commission { get; set; }
 
     /// <summary>
     /// Contract size for 1 trading lot in currencies and futures
     /// </summary>
-    double? ContractSize { get; set; }
+    decimal? ContractSize { get; set; }
 
     /// <summary>
     /// Tick size, i.e. minimum price change 
     /// </summary>
-    double? StepSize { get; set; }
+    decimal? StepSize { get; set; }
 
     /// <summary>
     /// Tick value, i.e. how much price change within one tick
     /// </summary>
-    double? StepValue { get; set; }
+    decimal? StepValue { get; set; }
 
     /// <summary>
     /// Aggregation period for the quotes
@@ -67,32 +67,32 @@ namespace Terminal.Core.Domains
     /// <summary>
     /// Long swap rate for keeping position overnight
     /// </summary>
-    public virtual double? SwapLong { get; set; }
+    public virtual decimal? SwapLong { get; set; }
 
     /// <summary>
     /// Short swap rate for keeping position overnight
     /// </summary>
-    public virtual double? SwapShort { get; set; }
+    public virtual decimal? SwapShort { get; set; }
 
     /// <summary>
     /// Commission
     /// </summary>
-    public virtual double? Commission { get; set; }
+    public virtual decimal? Commission { get; set; }
 
     /// <summary>
     /// Contract size for 1 trading lot in currencies and futures
     /// </summary>
-    public virtual double? ContractSize { get; set; }
+    public virtual decimal? ContractSize { get; set; }
 
     /// <summary>
     /// Tick size, i.e. minimum price change 
     /// </summary>
-    public virtual double? StepSize { get; set; }
+    public virtual decimal? StepSize { get; set; }
 
     /// <summary>
     /// Tick value, i.e. how much price change within one tick
     /// </summary>
-    public virtual double? StepValue { get; set; }
+    public virtual decimal? StepValue { get; set; }
 
     /// <summary>
     /// Aggregation period for the quotes
@@ -114,12 +114,12 @@ namespace Terminal.Core.Domains
     /// </summary>
     public Instrument()
     {
-      SwapLong = 0.0;
-      SwapShort = 0.0;
-      StepSize = 0.01;
-      StepValue = 0.01;
-      Commission = 0.0;
-      ContractSize = 1.0;
+      SwapLong = 0;
+      SwapShort = 0;
+      StepSize = 0.01m;
+      StepValue = 0.01m;
+      Commission = 0;
+      ContractSize = 1;
 
       Points = [];
       PointGroups = [];
