+++ Core/Models/Points/BarModel.cs
@@ -7,22 +7,22 @@ namespace Terminal.Core.Models
     /// <summary>
     /// Lowest price of the bar
     /// </summary>
-    public virtual double? Low { get; set; }
+    public virtual decimal? Low { get; set; }
 
     /// <summary>
     /// Highest price of the bar
     /// </summary>
-    public virtual double? High { get; set; }
+    public virtual decimal? High { get; set; }
 
     /// <summary>
     /// Open price of the bar
     /// </summary>
-    public virtual double? Open { get; set; }
+    public virtual decimal? Open { get; set; }
 
     /// <summary>
     /// Close price of the bar
     /// </summary>
-    public virtual double? Close { get; set; }
+    public virtual decimal? Close { get; set; }
 
     /// <summary>
     /// Clone
