+++ Gateways/Alpaca/Libs/Messages/JsonNewOrderAdvancedAttributes.cs
@@ -5,8 +5,8 @@ namespace Alpaca.Markets;
 public class JsonNewOrderAdvancedAttributes
 {
   [JsonPropertyName("limit_price")]
-  public double? LimitPrice { get; set; }
+  public decimal? LimitPrice { get; set; }
 
   [JsonPropertyName("stop_price")]
-  public double? StopPrice { get; set; }
+  public decimal? StopPrice { get; set; }
 }
