+++ Gateways/Ameritrade/Libs/Models/ChainModels.cs
@@ -19,43 +19,43 @@ namespace Terminal.Gateway.Ameritrade.Models
     public string ExchangeName { get; set; }
 
     [JsonPropertyName("bid")]
-    public double? Bid { get; set; }
+    public decimal? Bid { get; set; }
 
     [JsonPropertyName("ask")]
-    public double? Ask { get; set; }
+    public decimal? Ask { get; set; }
 
     [JsonPropertyName("last")]
-    public double? Last { get; set; }
+    public decimal? Last { get; set; }
 
     [JsonPropertyName("mark")]
-    public double? Mark { get; set; }
+    public decimal? Mark { get; set; }
 
     [JsonPropertyName("bidSize")]
-    public double? BidSize { get; set; }
+    public decimal? BidSize { get; set; }
 
     [JsonPropertyName("askSize")]
-    public double? AskSize { get; set; }
+    public decimal? AskSize { get; set; }
 
     [JsonPropertyName("bidAskSize")]
     public string BidAskSize { get; set; }
 
     [JsonPropertyName("lastSize")]
-    public double? LastSize { get; set; }
+    public decimal? LastSize { get; set; }
 
     [JsonPropertyName("highPrice")]
-    public double? HighPrice { get; set; }
+    public decimal? HighPrice { get; set; }
 
     [JsonPropertyName("lowPrice")]
-    public double? LowPrice { get; set; }
+    public decimal? LowPrice { get; set; }
 
     [JsonPropertyName("openPrice")]
-    public double? OpenPrice { get; set; }
+    public decimal? OpenPrice { get; set; }
 
     [JsonPropertyName("closePrice")]
-    public double? ClosePrice { get; set; }
+    public decimal? ClosePrice { get; set; }
 
     [JsonPropertyName("totalVolume")]
-    public double? TotalVolume { get; set; }
+    public decimal? TotalVolume { get; set; }
 
     [JsonPropertyName("tradeDate")]
     public DateTime TradeDate { get; set; }
@@ -67,40 +67,40 @@ namespace Terminal.Gateway.Ameritrade.Models
     public long? TradeTimeInLong { get; set; }
 
     [JsonPropertyName("netChange")]
-    public double? NetChange { get; set; }
+    public decimal? NetChange { get; set; }
 
     [JsonPropertyName("volatility")]
-    public double? Volatility { get; set; }
+    public decimal? Volatility { get; set; }
 
     [JsonPropertyName("delta")]
-    public double? Delta { get; set; }
+    public decimal? Delta { get; set; }
 
     [JsonPropertyName("gamma")]
-    public double? Gamma { get; set; }
+    public decimal? Gamma { get; set; }
 
     [JsonPropertyName("theta")]
-    public double? Theta { get; set; }
+    public decimal? Theta { get; set; }
 
     [JsonPropertyName("vega")]
-    public double? Vega { get; set; }
+    public decimal? Vega { get; set; }
 
     [JsonPropertyName("rho")]
-    public double? Rho { get; set; }
+    public decimal? Rho { get; set; }
 
     [JsonPropertyName("timeValue")]
-    public double? TimeValue { get; set; }
+    public decimal? TimeValue { get; set; }
 
     [JsonPropertyName("openInterest")]
-    public double? OpenInterest { get; set; }
+    public decimal? OpenInterest { get; set; }
 
     [JsonPropertyName("inTheMoney")]
     public bool? IsInTheMoney { get; set; }
 
     [JsonPropertyName("theoreticalOptionValue")]
-    public double? TheoreticalOptionValue { get; set; }
+    public decimal? TheoreticalOptionValue { get; set; }
 
     [JsonPropertyName("theoreticalVolatility")]
-    public double? TheoreticalVolatility { get; set; }
+    public decimal? TheoreticalVolatility { get; set; }
 
     [JsonPropertyName("mini")]
     public bool? IsMini { get; set; }
@@ -112,10 +112,10 @@ namespace Terminal.Gateway.Ameritrade.Models
     public List<OptionDeliverables> OptionDeliverablesList { get; set; }
 
     [JsonPropertyName("strikePrice")]
-    public double? StrikePrice { get; set; }
+    public decimal? StrikePrice { get; set; }
 
     [JsonPropertyName("lastTradingDay")]
-    public double? LastTradingDay { get; set; }
+    public decimal? LastTradingDay { get; set; }
 
     [JsonPropertyName("expirationDate")]
     public long? ExpirationDate { get; set; }
@@ -124,10 +124,10 @@ namespace Terminal.Gateway.Ameritrade.Models
     public string ExpirationType { get; set; }
 
     [JsonPropertyName("multiplier")]
-    public double? Multiplier { get; set; }
+    public decimal? Multiplier { get; set; }
 
     [JsonPropertyName("intrinsicValue")]
-    public double? IntrinsicValue { get; set; }
+    public decimal? IntrinsicValue { get; set; }
 
     [JsonPropertyName("settlementType")]
     public string SettlementType { get; set; }
@@ -142,13 +142,13 @@ namespace Terminal.Gateway.Ameritrade.Models
     public bool? IsPennyPilot { get; set; }
 
     [JsonPropertyName("percentChange")]
-    public double? PercentChange { get; set; }
+    public decimal? PercentChange { get; set; }
 
     [JsonPropertyName("markChange")]
-    public double? MarkChange { get; set; }
+    public decimal? MarkChange { get; set; }
 
     [JsonPropertyName("markPercentChange")]
-    public double? MarkPercentChange { get; set; }
+    public decimal? MarkPercentChange { get; set; }
   }
 
   public struct OptionDeliverables
@@ -181,7 +181,7 @@ namespace Terminal.Gateway.Ameritrade.Models
     public string Strategy { get; set; }
     
     [JsonPropertyName("interval")]
-    public double? Interval { get; set; }
+    public decimal? Interval { get; set; }
     
     [JsonPropertyName("isDelayed")]
     public bool? IsDelayed { get; set; }
@@ -190,16 +190,16 @@ namespace Terminal.Gateway.Ameritrade.Models
     public bool? IsIndex { get; set; }
     
     [JsonPropertyName("daysToExpiration")]
-    public double? DaysToExpiration { get; set; }
+    public decimal? DaysToExpiration { get; set; }
 
     [JsonPropertyName("interestRate")]
-    public double? InterestRate { get; set; }
+    public decimal? InterestRate { get; set; }
     
     [JsonPropertyName("underlyingPrice")]
-    public double? UnderlyingPrice { get; set; }
+    public decimal? UnderlyingPrice { get; set; }
     
     [JsonPropertyName("volatility")]
-    public double? Volatility { get; set; }
+    public decimal? Volatility { get; set; }
     
     [JsonPropertyName("callExpDateMap")]
     public Dictionary<string, Dictionary<string, List<Option>>> CallExpDateMap { get; set; }
@@ -211,22 +211,22 @@ namespace Terminal.Gateway.Ameritrade.Models
   public struct Underlying
   {    
     [JsonPropertyName("ask")]
-    public double? Ask { get; set; }
+    public decimal? Ask { get; set; }
     
     [JsonPropertyName("askSize")]
-    public double? AskSize { get; set; }
+    public decimal? AskSize { get; set; }
     
     [JsonPropertyName("bid")]
-    public double? Bid { get; set; }
+    public decimal? Bid { get; set; }
     
     [JsonPropertyName("bidSize")]
-    public double? BidSize { get; set; }
+    public decimal? BidSize { get; set; }
     
     [JsonPropertyName("change")]
-    public double? Change { get; set; }
+    public decimal? Change { get; set; }
     
     [JsonPropertyName("close")]
-    public double? Close { get; set; }
+    public decimal? Close { get; set; }
     
     [JsonPropertyName("delayed")]
     public bool? Delayed { get; set; }
@@ -238,45 +238,45 @@ namespace Terminal.Gateway.Ameritrade.Models
     public string ExchangeName { get; set; }
     
     [JsonPropertyName("fiftyTwoWeekHigh")]
-    public double? FiftyTwoWeekHigh { get; set; }
+    public decimal? FiftyTwoWeekHigh { get; set; }
     
     [JsonPropertyName("fiftyTwoWeekLow")]
-    public double? FiftyTwoWeekLow { get; set; }
+    public decimal? FiftyTwoWeekLow { get; set; }
     
     [JsonPropertyName("highPrice")]
-    public double? HighPrice { get; set; }
+    public decimal? HighPrice { get; set; }
     
     [JsonPropertyName("last")]
-    public double? Last { get; set; }
+    public decimal? Last { get; set; }
     
     [JsonPropertyName("lowPrice")]
-    public double? LowPrice { get; set; }
+    public decimal? LowPrice { get; set; }
     
     [JsonPropertyName("mark")]
-    public double? Mark { get; set; }
+    public decimal? Mark { get; set; }
     
     [JsonPropertyName("markChange")]
-    public double? MarkChange { get; set; }
+    public decimal? MarkChange { get; set; }
     
     [JsonPropertyName("markPercentChange")]
-    public double? MarkPercentChange { get; set; }
+    public decimal? MarkPercentChange { get; set; }
     
     [JsonPropertyName("openPrice")]
-    public double? OpenPrice { get; set; }
+    public decimal? OpenPrice { get; set; }
     
     [JsonPropertyName("percentChange")]
-    public double? PercentChange { get; set; }
+    public decimal? PercentChange { get; set; }
     
     [JsonPropertyName("quoteTime")]
-    public double? QuoteTime { get; set; }
+    public decimal? QuoteTime { get; set; }
     
     [JsonPropertyName("symbol")]
     public string Symbol { get; set; }
     
     [JsonPropertyName("totalVolume")]
-    public double? TotalVolume { get; set; }
+    public decimal? TotalVolume { get; set; }
     
     [JsonPropertyName("tradeTime")]
-    public double? TradeTime { get; set; }
+    public decimal? TradeTime { get; set; }
   }
 }
